# docs/openapi.json
{
  "components": {
    "schemas": {
      "AlertRequest": {
        "additionalProperties": false,
        "description": "A message that takes the whole sign over until it is released.",
        "properties": {
          "display_mode": {
            "default": "HOLD",
            "description": "how the sign presents the alert",
            "title": "Display Mode",
            "type": "string"
          },
          "message": {
            "description": "the alert text. The sign's priority file holds 125 bytes once markup has been rendered, and cannot be resized. It cannot be empty: a write with no text still carries the formatting bytes around it, which the sign reads as a blank priority message and displays, so the sign would sit blank with the rotation suppressed behind it and an alert reported as active",
            "maxLength": 4096,
            "minLength": 1,
            "title": "Message",
            "type": "string"
          },
          "ttl_seconds": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "release the sign this many seconds from now. Omit and the alert holds the sign until something releases it explicitly",
            "title": "Ttl Seconds"
          }
        },
        "required": [
          "message"
        ],
        "title": "AlertRequest",
        "type": "object"
      },
      "AlertResponse": {
        "description": "The alert currently holding the sign.",
        "properties": {
          "display_mode": {
            "title": "Display Mode",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          }
        },
        "required": [
          "message",
          "display_mode",
          "started_at",
          "expires_at"
        ],
        "title": "AlertResponse",
        "type": "object"
      },
      "ClockResponse": {
        "description": "The result of setting the sign's clock.",
        "properties": {
          "synced_at": {
            "description": "when the clock was set, in its configured zone. Not what the sign was told: the sign is deliberately set one minute ahead, because the protocol's Set Time carries no seconds and a sign told the current minute runs behind for the rest of it",
            "format": "date-time",
            "title": "Synced At",
            "type": "string"
          }
        },
        "required": [
          "synced_at"
        ],
        "title": "ClockResponse",
        "type": "object"
      },
      "ControlCommandRequest": {
        "additionalProperties": false,
        "description": "A command aimed at the sign itself rather than at a message.",
        "properties": {
          "command": {
            "description": "one of SET_DAY_OF_WEEK, SET_TIME, SET_TIME_FORMAT, SOFT_RESET, SOUND, SPEAKER",
            "title": "Command",
            "type": "string"
          },
          "parameter": {
            "default": "",
            "description": "the command's parameter, if it takes one",
            "title": "Parameter",
            "type": "string"
          }
        },
        "required": [
          "command"
        ],
        "title": "ControlCommandRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SlotResponse": {
        "description": "A registered slot.",
        "properties": {
          "display_mode": {
            "title": "Display Mode",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "description": "the sign file this slot occupies, A through Z",
            "title": "Label",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "order": {
            "title": "Order",
            "type": "integer"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "key",
          "label",
          "message",
          "display_mode",
          "order",
          "source",
          "expires_at",
          "updated_at"
        ],
        "title": "SlotResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "description": "The shared key every write carries. Set it with `api_key` in the config file or `READERBOARD_API_KEY` in the environment; `scripts/install.sh` generates one.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      }
    }
  },
  "info": {
...
    "title": "readerboard",
    "version": "0.4.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/alerts": {
      "delete": {
        "description": "Release an alert so the rotation resumes.",
        "operationId": "delete_alert_alerts_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Give the sign back",
        "tags": [
          "Alerts"
        ]
      },
      "get": {
        "description": "Return the active alert, or null if the sign is rotating normally.",
        "operationId": "get_alert_alerts_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/AlertResponse"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Get Alert Alerts Get"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Read the alert holding the sign",
        "tags": [
          "Alerts"
        ]
      },
      "post": {
        "description": "Take the whole sign over until the alert is released.\n\nThis uses the sign's priority file, which suppresses every other message. If\na ttl is given, the sign is released automatically and the rotation resumes\nby itself.",
        "operationId": "post_alert_alerts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Take the sign over with an alert",
        "tags": [
          "Alerts"
        ]
      }
    },
...
    "/messages": {
      "delete": {
        "description": "Remove every slot, leaving the sign showing nothing.",
        "operationId": "clear_messages_messages_delete",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Take every message off the sign",
        "tags": [
          "Messages"
        ]
      },
      "get": {
        "description": "Return every registered slot, in the order the sign plays them.",
        "operationId": "list_messages_messages_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SlotResponse"
                  },
                  "title": "Response List Messages Messages Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List the messages sharing the sign",
        "tags": [
          "Messages"
        ]
      }
    },
    "/messages/{key}": {
      "delete": {
        "description": "Remove one slot and free the sign file it held.",
        "operationId": "delete_message_messages__key__delete",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "description": "the name of the slot, chosen by whoever owns it",
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9._-]+$",
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Take a message off the sign",
        "tags": [
          "Messages"
        ]
      },
      "get": {
        "description": "Return one slot by name.",
        "operationId": "get_message_messages__key__get",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "description": "the name of the slot, chosen by whoever owns it",
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9._-]+$",
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read one message",
        "tags": [
          "Messages"
        ]
      },
      "put": {
        "description": "Put a message in a slot, replacing whatever was there.\n\nThe sign rotates through every registered slot on its own, so registering a\nsecond message does not displace the first.",
        "operationId": "put_message_messages__key__put",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "description": "the name of the slot, chosen by whoever owns it",
              "maxLength": 64,
              "minLength": 1,
              "pattern": "^[A-Za-z0-9._-]+$",
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Register or replace a message",
        "tags": [
          "Messages"
        ]
      }
    },
    "/sign/command": {
      "post": {
        "description": "Send one of the sign's own control commands.\n\n`SOFT_RESET` restarts the sign. It erases nothing, the sign comes back\nshowing what it was showing, and it is the first thing to try on a sign that\nhas stopped responding. The display is blank for a few seconds while it runs\nits power-up diagnostics, and this call waits that out before answering, so\na 204 means the sign is listening again. Reach for `POST /sign/reboot` only\nwhen a soft reset is not enough: that one erases the sign and rebuilds it.",
        "operationId": "send_command_sign_command_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlCommandRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Send a control command to the sign",
        "tags": [
          "Sign"
        ]
      }
    },
    "/sign/information": {
      "get": {
        "description": "Read the sign's own account of itself.\n\nThe firmware build and the month it was released, the sign's clock and\nwhether it draws a 12 or 24 hour one, whether its speaker is enabled, and\nhow much of its memory pool is free. The protocol document calls this \"most\nuseful as a source of troubleshooting information\", which is a fair summary:\nnothing here changes anything.\n\nThis is the only read in the service, so it is also the only place a silent\nsign is distinguishable from an unplugged one. A sign that does not answer\nwithin a few seconds is a 503, the same as a sign that cannot be written to,\nand so is a sign that answers with something this cannot read: the reply is\nthe whole of what the endpoint has, so one that will not parse leaves it\nwith nothing to report.\n\nTwo fields are worth reading carefully. `speaker_enabled` is why `SOUND` can\nappear to do nothing: the protocol calls disabled the default, and a muted\nsign beeps silently. And `memory_free` is the pool the memory configuration\ndraws on, so a slot capacity that will not fit shows up here before it shows\nup as a failed reallocation.",
        "operationId": "sign_information_sign_information_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignInformationResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Ask the sign what it is and how it is doing",
        "tags": [
          "Sign"
        ]
      }
    },
    "/sign/reboot": {
      "post": {
        "description": "Reset the sign to recover it, then put everything back.\n\n**A recovery tool, not a routine operation, and it is disruptive.** It sends\nthe memory clear that resets the sign, which erases every file on it and\ntakes the display blank for several seconds while the sign restarts. The\nservice then waits for the reset to finish and re-pushes every message and\nthe run sequence from its own record, so the sign comes back showing what it\nwas showing before rather than empty. Any active alert is re-asserted as\npart of that restore.\n\nUse it for a sign that has stopped responding to writes or is showing\ngarbage, the wedged-decoder state a unit mounted out of reach can fall into\nfrom a stray bit and that cannot be fixed by power cycling it by hand. Do\nnot use it to clear the sign: `DELETE /messages` takes every message off\nwithout resetting anything, and this puts them all straight back. Expect a\nblank display for roughly ten seconds before the rotation returns.\n\n503 if the sign cannot be reached, since a sign that is not answering cannot\nbe rebooted.",
        "operationId": "reboot_sign_sign_reboot_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Reboot the sign to recover it, then restore the display",
        "tags": [
          "Sign"
        ]
      }
    },
    "/sign/sync-clock": {
      "post": {
        "description": "Set the sign's clock and day of week immediately.\n\nThe service already does this at startup, hourly, and whenever the link to\nthe sign comes back. This is for when you would rather not wait.\n\nThe sign is set one minute ahead of the real time, on purpose. Set Time\ncarries four digits and no seconds, so a sign told the current minute starts\nthat minute over and then reads behind for the rest of it, by up to\nfifty-nine seconds, and never ahead. Leading by a minute moves that\none-sided error to the side where the sign is at worst a minute fast and\nnever slow.",
        "operationId": "sync_clock_sign_sync_clock_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClockResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Set the sign's clock now",
        "tags": [
          "Sign"
        ]
      }
    },
...
  }
}